reduce duplicate code in card value objects
reported by sonar analysis
